Addition of service tests

+++ src/main/java/myretail/products/api/domain/CurrentPrice.java
@@ -5,13 +5,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Document(value = "current_price")
 public class CurrentPrice {
-    float value;
+    BigDecimal value;
     @Field("currency_code")
     String currencyCode;
 }
